requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
slowapi==0.1.8
openai==1.3.5
spacy==3.7.2
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==0.8.11
pdf2image==1.17.0
httpx==0.25.2
redis==5.0.1
python-dateutil==2.8.2
scikit-learn==1.3.2
numpy==1.26.2
pytesseract==0.3.10
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
